+++ WebSites/WCFDiagnostikalNexus/App_Code/BLL/parser.cs
@@ -25,7 +25,7 @@ public class parser
             MD5 md5 = MD5CryptoServiceProvider.Create();
             byte[] dataMd5 = md5.ComputeHash(Encoding.Default.GetBytes(dato));
             StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < dataMd5.Length; i++)
+            for (long i = 0; i < dataMd5.Length; i++)
                 sb.AppendFormat("{0:x2}", dataMd5[i]);
             return sb.ToString();
